Use native Thor methods to add gems
I wasn't aware that Thor had methods to put gems in groups, as opposed to inline groups. Because of this, I was able to delete a decent chunk of code. It does have the caveat of making the gemfile much messier because it doesn't put gems in alphabetical order, and also make additional groups of the same name.

I decided that this was okay. Even with my custom method, there was some manual intervention needed. This way, at least my methods are supported officially.